+++ Team_Metrics/Offensive_Rating.py
@@ -83,13 +83,13 @@ def offensive_rating_get_data_set(match_data : dict={}) -> list:
         home_team_stats["third_period_goals"]
     try:
         power_play_data[home_team] = [
-            float(home_team_stats["power_play_goals"]),
-            float(home_team_stats["power_play_chances"])
+            home_team_stats["power_play_goals"],
+            home_team_stats["power_play_chances"]
         ]
     except KeyError:
         power_play_data[home_team] = [
-            float(home_team_stats["power_play_goals"]),
-            float(home_team_stats["penalty_minutes"] / 2)
+            home_team_stats["power_play_goals"],
+            home_team_stats["penalty_minutes"] / 2
         ]
     
     # away data
@@ -99,13 +99,13 @@ def offensive_rating_get_data_set(match_data : dict={}) -> list:
         away_team_stats["third_period_goals"]
     try:
         power_play_data[away_team] = [
-            float(away_team_stats["power_play_goals"]),
-            float(away_team_stats["power_play_chances"])
+            away_team_stats["power_play_goals"],
+            away_team_stats["power_play_chances"]
         ]
     except KeyError:
         power_play_data[away_team] = [
-            float(away_team_stats["power_play_goals"]),
-            float(away_team_stats["penalty_minutes"] / 2)
+            away_team_stats["power_play_goals"],
+            away_team_stats["penalty_minutes"] / 2
         ]
     return {
         'shots_for' : shots_for_data,
